+++ changelog.md
@@ -3,6 +3,7 @@
 ## 0.3.0
 - Access component instance's name with `vm.$name` (rather than `vm.$options.name`)  
 - Access component's html content with `vm.$html` (rather than `vm.$el.outerHTML`)
+- `vm.$find` allows you to search a component's html for matching elements or component instances.
 
 ## 0.2.0
 - stubComponents flag automatically stubs all *known* components